#63 move bc implementation to FiniteVolumeDiscretisation (remove MatrixVectorDiscretisation class)
