+++ docs/source/discretisations/base_discretisation.rst
@@ -1,0 +1,5 @@
+Base Discretisation
+===================
+
+.. autoclass:: pybamm.BaseDiscretisation
+  :members:
