--- docs/source/discretisations/core.rst
@@ -1,8 +1,0 @@
-Core Discretisation
-===================
-
-.. autoclass:: pybamm.BaseDiscretisation
-  :members:
-
-.. autoclass:: pybamm.MatrixVectorDiscretisation
-  :members:
