+++ pybamm/discretisations/base_discretisation.py
@@ -233,82 +233,3 @@ class BaseDiscretisation(object):
 
     def concatenate(self, *symbols):
         return pybamm.NumpyConcatenation(*symbols)
-
-
-class MatrixVectorDiscretisation(BaseDiscretisation):
-    """
-    A base class for discretisations where the gradient and divergence operators are
-    always matrix-vector multiplications.
-
-    Parameters
-    ----------
-     mesh : :class:`pybamm.BaseMesh` (or subclass)
-            The underlying mesh for discretisation
-
-    **Extends:** :class:`pybamm.BaseDiscretisation`
-    """
-
-    def __init__(self, mesh):
-        super().__init__(mesh)
-
-    def gradient(self, symbol, domain, y_slices, boundary_conditions):
-        """Matrix-vector multiplication to implement the gradient operator.
-        See :meth:`pybamm.BaseDiscretisation.gradient`
-        """
-        # Check that boundary condition keys are hashes (ids)
-        for key in boundary_conditions.keys():
-            assert isinstance(key, int), TypeError(
-                "boundary condition keys should be hashes, not {}".format(type(key))
-            )
-        # Discretise symbol
-        discretised_symbol = self.process_symbol(
-            symbol, domain, y_slices, boundary_conditions
-        )
-        # Add boundary conditions if defined
-        if symbol.id in boundary_conditions:
-            lbc = boundary_conditions[symbol.id]["left"]
-            rbc = boundary_conditions[symbol.id]["right"]
-            discretised_symbol = self.concatenate(lbc, discretised_symbol, rbc)
-        gradient_matrix = self.gradient_matrix(domain)
-        return gradient_matrix * discretised_symbol
-
-    def gradient_matrix(self, domain):
-        """The gradient matrix
-
-        Parameters
-        ----------
-        domain : list
-            The domain(s) in which to compute the gradient matrix
-        """
-        raise NotImplementedError
-
-    def divergence(self, symbol, domain, y_slices, boundary_conditions):
-        """Matrix-vector multiplication to implement the divergence operator.
-        See :meth:`pybamm.BaseDiscretisation.gradient`
-        """
-        # Check that boundary condition keys are hashes (ids)
-        for key in boundary_conditions.keys():
-            assert isinstance(key, int), TypeError(
-                "boundary condition keys should be hashes, not {}".format(type(key))
-            )
-        # Discretise symbol
-        discretised_symbol = self.process_symbol(
-            symbol, domain, y_slices, boundary_conditions
-        )
-        # Add boundary conditions if defined
-        if symbol.id in boundary_conditions:
-            lbc = boundary_conditions[symbol.id]["left"]
-            rbc = boundary_conditions[symbol.id]["right"]
-            discretised_symbol = self.concatenate(lbc, discretised_symbol, rbc)
-        divergence_matrix = self.divergence_matrix(domain)
-        return divergence_matrix * discretised_symbol
-
-    def divergence_matrix(self, domain):
-        """The divergence matrix
-
-        Parameters
-        ----------
-        domain : list
-            The domain(s) in which to compute the divergence matrix
-        """
-        raise NotImplementedError
